Add timeout to Amazon's getProducts and getUserData requests (#630)
### Description
https://revenuecats.atlassian.net/browse/CSDK-452

In certain situations, Amazon fails silently and doesn't return anything
for the requests. For example, when enabling [sandbox mode
](https://developer.amazon.com/docs/in-app-purchasing/iap-app-tester-user-guide.html#installtester)
and using the version from Live App testing. In those situations, we
want to add a timeout to the requests so it doesn't get stuck and we
show a meaningful error log. I added it to both the `getUserData` and
`getProducts` requests.