[CF-971] BC5 postReceipt updates (#698)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [will be in separate PR] If applicable, unit tests
- [n/a, will follow up for hybrids once api is finalized] If applicable,
create follow-up issues for `purchases-ios` and hybrids

### Motivation
<!-- Why is this change required? What problem does it solve? -->
In BillingClient5, we do not receive price information in the `Purchase`
object returned when a purchase completes. Instead, we need to store
which `PurchaseOption` the purchase was initiated for, pass the
associated `PricingPhase`s to our backend, and let our backend handle
the price reporting.

For restored purchases, we are not in control of the purchase flow and
therefore don't know the `PurchaseOption` which triggered the purchase.
In those cases, our backend will have to use the Google APIs to deduce
the price associated with the Purchase.

<!-- Please link to issues following this format: Resolves #999999 -->
Resolves https://revenuecats.atlassian.net/browse/CF-887

### Description
<!-- Describe your changes in detail -->
Store the `purchaseOptionId` associated with the `Purchase` and the
associated purchase callback, similar to how we previously stored the
`sku`. Then, when a purchase completes, pass the `PricingPhase`s of that
`PurchaseOption` to the backend.

I also removed a lot of the fields we no longer need to post to the
backend.

<!-- Please describe in detail how you tested your changes -->
I ran through a purchase flow in the sample app and confirmed that
postReceipt returned 200. Unit tests will be a in a separate PR, and
end-to-end testing in the bug bash and with the testing restores ticket.