Disable unsupported features for first BC5 beta (#701)
### Description
https://revenuecats.atlassian.net/browse/CF-989

For the first BC5 beta, not all features will be available. We will
document this for our beta users but additionally, it's good to try to
fail early if beta users try to use unsupported features. This PR adds
some checks so we fail as early as possible when using any of these
unsupported features.