Refactor postReceipt helper method (#703)
We had a postReceipt() method called in `BackendTest` that I found a
little confusing to work with. It calls both mockPostReceiptResponse and
then posts to the backend. We had duplicate code mocking/posting for the
tests using the asyncBackend.

This PR:
* renames postReceipt to make more clear that it is a test function
* update it to take a backend parameter so it can be reused for
asyncBackend calls
* removes the fetchToken as a return value -- it was just hard-coded
every time anyway
* update stubStoreProduct to not take in duration, since that's a
pricingphase idea


tests for postReceipt will still be in a different PR, this was just
some clean up to make those easier to add