Fix PostingTransactionsTests (#711)
Fixes failing tests in `PostingTransactionsTests`

One interesting find -- you cannot `capture` a nullable type. This was
leading to some of the failures...

we had
```
val postedStoreUserIdSlot = slot<String>()
every {
            backendMock.postReceiptData(
                ...
                storeAppUserID = capture(postedStoreUserIdSlot),
                ...
            )
} answers { ... }
```

and when the storeUserId was null, it wasn't matching the calls. you
can't make a `slot<String?>`, so the
[recommendation](https://github.com/mockk/mockk/issues/293) for
nullables is apparently to use a list, like i did here. idk if that's a
well-known thing, but it definitely caught me..