Default ProrationMode to IMMEDIATE_WITHOUT_PRORATION (#740)
BC4 used to default to `IMMEDIATE_WITH_TIME_PRORATION` when no
`ProrationMode` was passed in, but BC5 will fail with no
`ProrationMode`, so we now need a default.

Until we complete other ProrationMode logic, the best default is
`IMMEDIATE_WITHOUT_PRORATION`. This will work for both base plan and sub
switches.