[CF-1135] Remove requirement for `StoreProduct` when purchasing a `PurchaseOption` (#739)
## Motivation

[CF-1135](https://revenuecats.atlassian.net/browse/CF-1135)

The following methods required a `StoreProduct` when purchasing a
`PurchaseOption`. A `PurchaseOption` belongs to a `StoreProduct` so this
was duplicate information that developers should not have to pass in.

```kotlin
fun purchaseSubscriptionOption(
    activity: Activity,
    storeProduct: StoreProduct,
    purchaseOption: PurchaseOption,
    upgradeInfo: UpgradeInfo,
    listener: ProductChangeCallback
) 

fun purchaseSubscriptionOption(
    activity: Activity,
    storeProduct: StoreProduct,
    purchaseOption: PurchaseOption,
    callback: PurchaseCallback
) 
```

The goal was to only require `purchaseOption: PurchaseOption` to make a
simpler API and prevent developer passing in a `StoreProduct` and
`PurchaseOption` that weren't related to each other.

## Description

### Updated Public APIs

```kotlin
fun purchaseSubscriptionOption(
    activity: Activity,
    purchaseOption: PurchaseOption,
    upgradeInfo: UpgradeInfo,
    listener: ProductChangeCallback
) 

fun purchaseSubscriptionOption(
    activity: Activity,
    purchaseOption: PurchaseOption,
    callback: PurchaseCallback
) 
```

### New BillingWrapper Architecture

Previously, `BillingAbstract.makePurchaseAsync` required a
`StoreProduct` and `PurchaseOption?`. This was a big problem for this
change and also something would make Amazon support weird because Amazon
doesn't have a concept of `PurchaseOption`

`BillingAbstract.makePurchaseAsync` takes only a new `PurchasingData`
object:

```kotlin
abstract fun makePurchaseAsync(
    activity: Activity,
    appUserID: String,
    purchasingData: PurchasingData,
    replaceProductInfo: ReplaceProductInfo?,
    presentedOfferingIdentifier: String?
)
```

`makePurchaseAsync` is now fully abstracted away from the individual
store implementations again:
- `interface PurchasingData`
  - `class AmazonPurchasingData`
  - `class GooglePurchasingData`

`BillingWrapper` (Google Play implementation) and `AmazonBilling`
(Amazon implementation) now can require any and all information they
need in their respective concrete `PurchasingData` implementations
without needing to ever change the shared interface for
`BillingAbstract`

#### `GooglePurchasingData`

Google Play has two very distinct product types for subscriptions and
non-subscriptions so these are differentiated in a sealed
`GooglePurchasingData` class with two data classes: `NotSubscription`
and `Subscription`

The big difference here is that `Subscription` has a `token: String` and
`optionid: String?`.

```kotlin
sealed class GooglePurchasingData: PurchasingData {
    data class NotSubscription(
        override val productId: String,
        val productDetails: ProductDetails
    ) : GooglePurchasingData()

    data class Subscription(
        override val productId: String,
        val productDetails: ProductDetails,

        val optionId: String,
        val token: String
    ) : GooglePurchasingData()
}
```

This was **not** the only way to design the `GooglePurchasingData`. This
could have just been one data class with an `token: String?` and
`optionId: String?` but using a `sealed class` allows us to use a nice
`when` to switch on the types in `BillingClient` when needed. It also is
nice in the instantiation of `Subscription` because it insures that a
token exists and could prevent any bugs where it doesn't by giving us
compile errors.

#### `AmazonPurchasingData`

Due to the focus being Google right now, there were not thoughtful
changes on the Amazon side. The goal was to make it compile with as few
changes as possible. So for this, The `AmazonPurchasingData` just takes
an `AmazonStoreProduct`. Ideally this could be simplified to a
`productId: String` only (I think) but we'll save that for a different
day.

```kotlin
sealed class AmazonPurchasingData: PurchasingData {
    data class Product(
        val storeProduct: AmazonStoreProduct,
    )
}
```


### The Magic

- `StoreProduct` maps to `PurchasingData`
- `PurchaseOption` maps to `PurchasingData`

This is what allows any `Purchases.shared.purchase` method that takes
either of those to fully make a purchase by converting either a
`StoreProduct` or `PurchaseOption` to `PurchasingData` and passes that
to the implementation of `AbstractBilling`.

[CF-1135]:
https://revenuecats.atlassian.net/browse/CF-1135?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ