[CF-1119] - Default offer now uses best offer logic (longest free, best savings, base plan) (#754)
## Checklist
- [x] If applicable, unit tests
- [x] If applicable, create follow-up issues for `purchases-ios` and
hybrids

## Motivation

[CF-1119](https://revenuecats.atlassian.net/browse/CF-1119)

The previous behavior purchasing a package or product was to use the
base plan. That is the full priced product and won't use any of the
offers (free trials or discounts) that the developer has setup. This
change is to perform logic to choose the "best offer" for the developer.

The order is:
1. Longest free trial
2. Cheapest price
3. Fall back to base plan

<img width="543" alt="Screen Shot 2023-01-25 at 10 54 45 AM"
src="https://user-images.githubusercontent.com/401294/214628814-a4b3676b-9909-41f3-a7e8-5881688f3c18.png">


## Description

- New `findBestOffer.kt` file that adds
`List<GoogleSubscriptionOption>.findBestOffer():
GoogleSubscriptionOption?`
- `findBestOffer()` is called in `ProductDetails.toStoreProduct()`
  - This is where the `defaultOption` was previous the base plan

### How `findBestOffer()` works

1. Returns `null` if no base plan in list of `SubscriptionOption`s
    - Its needed because its the fallback
    - Its needed to find base savings
    - It _should_ always be there for a subscription
2. Filters out base plan and offers tagged with `rc-ignore-best-offer`
in `SubscriptionOption`s
- Base plan is not a thing that can be used in calculating best off (its
the fallback)
    - Removes offers that developer explicitly don't want to be used
3. Finds and returns longest free trial with `findLongestFreeTrial()`
(if exists)
5. Finds and returns cheapest offer with `findLowestNoneFreeOffer()` (if
exists)
7. Falls back to returning base plan

#### Why custom ISO 8601 duration parsing?

Kotlin has a `Duration.parse` function (see
https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.time/-duration/parse.html)
but its only available in API 26 or higher. We can't bump our min so
decision was to make an internal implementation 🤷‍♂️

Uses a regex to parse the year, month, week, and day components. These
numbers are then used to get a rough number of total days for the
duration to help calculate longest free trial and biggest savings

[CF-1119]:
https://revenuecats.atlassian.net/browse/CF-1119?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Maddie Beyl <maddie@revenuecat.com>