Diagnostics: Diagnostics split manager into tracker and synchronizer (#795)
### Description
Based on #793 
This PR refactors previous code to split the `DiagnosticsManager` into
`DiagnosticsTracker` and `DiagnosticsSynchronizer`. This is a
prerequisite to avoid circular dependencies when tracking request data.
- Before:
`DiagnosticsManager`->`Backend`->`HttpClient`->`DiagnosticsManager`.
- After: `DiagnosticsSynchronizer`->`Backend` for syncing and
`HttpClient`->`DiagnosticsTracker` for tracking requests

I could have tried to extract this logic from the HttpClient, but that
would be a bigger refactor. This was the simplest solution I thought for
this problem. Also, sorry I didn't see that issue before. I didn't
modify the #785 PR directly to try to avoid some conflicts that would
happen if I did it there directly.