Entitlement verification: Architecture and API (#820)
### Description
This PR has the initial changes to support entitlement verification in
the Android SDK including the basic architecture and wiring.
I worked on several things so this got bigger than I wanted, sorry about
that! Most of the changes are on tests though.

### Changes:
- Add nonce header to request when entitlement verification enabled and
endpoint supports it.
- Add `VerificationStatus` to `HTTPResult` indicating whether the
verification happened and whether it was successful or not.
- Error request when VerificationMode is `ENFORCED` and signature
verification fails.
- Add public API to enable/disable entitlement verification. Current
supported modes:
   - `DISABLED`: No verification will be performed
- `INFORMATIONAL`: Verification will be performed but entitlements will
still be granted on error. We will log an error and put this state in
the `EntitlementInfo` in a future PR.
- `ENFORCED`: Verification will be performed and cause the request to
fail on verification error, so no entitlements will be granted.


### Remaining tasks (in followup PRs):
- [ ] Add verification library and perform actual verification
- [ ] Add verification status to EntitlementInfo
- [ ] Update response cache with request time