Entitlement verification: Return current call verification result over cached version (#842)
### Description
This PR makes sure we expose the current call verification result over
the cached version. We shouldn't be storing verification errors in the
etag cache. Also, we plan to clear the cache when verification mode is
enabled, so we shouldn't have `NOT_REQUESTED` values either, so the etag
cache should only contain successful values