Merge main ->  bc5-support 2023-03-15 (#875)
### Description
This brings latest `main` to `bc5-support`. The changes include
diagnostics and trusted entitlements.

The changes tracking google methods with diagnostics were simply removed
and will be added in a followup PR, since the original changes were only
done in the BC4 version of the `BillingWrapper` and it's easier to just
reimplement that part.

There were a bunch of conflicts, but most were easy to solve, except
some in the unit tests. Been doing a pass to make sure I didn't forget
any.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Cesar de la Vega <cesarvegaro@gmail.com>
Co-authored-by: RevenueCat Git Bot <72824662+RCGitBot@users.noreply.github.com>
Co-authored-by: revenuecat-ops <ops@revenuecat.com>
Co-authored-by: Andy Boedo <andresboedo@gmail.com>
Co-authored-by: AristiDevs <aris_ks@hotmail.com>
Co-authored-by: beylmk <maddie@revenuecat.com>
Co-authored-by: NachoSoto <NachoSoto@users.noreply.github.com>
Co-authored-by: NachoSoto <ignaciosoto90@gmail.com>
Co-authored-by: Josh Holtz <me@joshholtz.com>
Co-authored-by: HaleyRevcat <112121235+HaleyRevcat@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Stefan Wehner <stefan@revenuecat.com>