Hide all new Trusted entitlements APIs (#886)
### Description
This completes SDK-3000

We aren't ready to ship trusted entitlements just yet. Until then, we
are hiding the public APIs so it remains disabled.

This was trickier than originally expected due to our multi-module
setup, so we couldn't just make the API `internal` since much of our
public API lives in the `public` module and needs to be accessible by
the `common` module.

We could refactor things to have a different public API and internal
API, which might be something we should do moving forward... But
currently our public API is used extensively in our internal code.