Offline entitlements: Add `OfflineEntitlementsManager` and add operation to perform backend request and cache result in device (#893)
### Description
Second part of SDK-2986
Based on #892 

This PR adds `OfflineEntitlementsManager`. This class will be in charge
of performing most of the logic to make sure the SDK can handle
entitlements in the unlikely case the servers are down.

Currently, only has one method to update the cache with the product
entitlement mappings from the servers.