Add integration tests that can mock the Store (#899)
### Description
We've been limited in our integration tests due to Google/Amazon being
very hard to test automatically. This PR takes an approach to mock the
store itself, so we can test the rest of the system. The changes in this
PR are:
- Add a new flavor to the `purchases` module: `integrationTest`
- This new flavor has code that allows injection of a custom
`BillingWrapper` and execute as instrumentation tests, like an Activity,
manifest, and icons. This won't be added to the normal SDK.
- Moved integration tests to the `purchases` module in the
`integrationTest` flavor. (I kept the old integration tests for now
until we have tested the new flow more, they can live together for now).
- Added tests that verify that `getOfferings` and `purchaseProduct`
work. We need to pass in an API key and a google purchase token

#### Next steps (In followup PRs)
- [ ] Automate new integration tests (need to be sent to firebase labs
to run)
- [ ] Add tests for product entitlement mappings
- [ ] Add tests for load shedder (which will probably be very close to
these).