Add load shedder integration tests to V4 version (#933)
### Description
We recently created the new load shedder server to take charge in case
the main servers fail. This mechanism is a failsafe to make sure we
don't lose any data and most customers aren't affected by this downtime.
In order to ensure that these emergency system works correctly when we
need it, we plan to run periodic tests that hit the load shedder and
make sure the SDKs keep working as expected.

This PR adds integration tests for the load shedder for SDK version 4.
Most of the changes are based off #899 and #911

This is pointing to branch
`v4-load-shedder-integration-tests-do-not-remove`. This branch will
remain open with the changes needed to run these integration tests
against the load shedder. And we will run them periodically from that
branch, but they won't get deployed in an SDK version.

#### TODO (in other PRs)
- [x] Add automation #934 
- [x] Add daily schedule to run these tests #934 
- [ ] Add slack notifications