Add back integration test flavors (#962)
### Description
We detected that the chnages to support the new integration tests caused
compilation issues for top-level kotlin functions. After debugging the
issue, we tracked the issue down to this block of code:
```
    productFlavors {
        integrationTest {
            .....
            packagingOptions {
                resources.excludes.add("META-INF/*")
            }
        }
    }
```
We assumed that the `packagingOptions` there would only affect the
`integrationTest` flavor, but it actually affects other flavors as well.
This was causing a `kotlin_module` file to not be included in the
published `aar` file, which caused top level functions to not be
accessible by library consumers.
In this PR we change the exclusion to only exclude the files that were
duplicated during instrumentation tests: `META-INF/LICENSE.md` and
`META-INF/LICENSE-notice.md`.
Comparing the resulting `AAR`'s:
- In 5.8.2: We had a
`META-INF/purchases_latestDependenciesRelease.kotlin_module` file.
- In 6.1.0 (the faulty version): We didn't have any `kotlin_module`
files
- In 6.1.1: We had a `META_INF/purchases_release.kotlin_module` file.
- After this fix: We have a
`META_INF/purchases_defaultsRelease.kotlin_module` file.

Also, testing this change (by deploying to local maven and using it in a
project) we can confirm the issue appears in 6.1.0 and before this
change but doesn't appear after this change.

#### TODO
- [ ] Add tests that check against the resulting aar instead of using
the modules to make sure this doesn't happen again. Added task SDK-3062
to track this