Extract post receipt logic to PostReceiptHelper (#967)
### Description
This became a complex refactor, specially on the tests... Basically, we
currently have 3 places where we post receipts to the backend: restores,
sync purchases and purchases. All this logic lived in `Purchases` and
part of it was duplicated. In this refactor, I'm extracting this logic
from all 3 places to a new `PostReceiptHelper` so we can share the
common logic between them.

The main code itself wasn't too bad to refactor, but the tests were very
painful to untangle, since we had them in different places.

The purpose of doing this now is so we can share the logic for offline
entitlements in #964 and also some very needed cleanup of `Purchases`.

#### TODO
- [x] Verify all test cases are also tested with the new code
- [x] Test this change deeply