Support enabling/disabling offline entitlements (#964)
### Description
This PR enables offline entitlements automatically when we receive 5xx
response codes in the `postReceipt` requests or on `getCustomerInfo` if
there isn't a cached version already. It also disables offline
entitlements when we receive a successful response in one of the above
requests or logIn/logout.

We went with a stateful approach since we need to return the offline
computed customer info from `getCustomerInfo` after a `postReceipt`
fails.

All this is also behind a flag right now that is hardcoded so we don't
enter offline entitlements mode yet.