Clear cached customer info upon entering offline entitlements mode (#989)
### Description
SDK-3095

Offline entitlements mode caches the computed customer info in memory.
That means it will be lost upon a restart. So if users restart, they
would temporarily go back to the cached customer info until we either
get a new value from the backend or it fails again and we enter offline
entitlement mode again. This can make it so developers see a strange
change in customer infos upon a restart after entering offline customer
info.

In order to solve this, we are clearing the cached customer info upon
entering offline entitlements mode in this PR. This way, next time the
user restarts the app, it will have to first try to get the value from
the backend, waiting to return a value, avoiding the changes in customer
info.