Add backend integration tests and test product entitlement mapping endpoint (#988)
### Description
SDK-3015

This PR adds backend integration tests. These are setup as unit tests in
the `common` module.

I thought about adding these as instrumentation tests, but that
increases the complexity and runtime of these tests, and we can
relatively simply run them as unit tests in the JVM.

This is an almost duplicate of #979. Those changes were lost after being
merged to the wrong branch. Only change after that is fixing the load
shedder test now that we are actually requesting to the load shedder
after some backend changes.