Add completion callback to syncPurchases API (#1002)
### Description
This PR adds a new optional parameter to the `syncPurchases` function
containing a success and error callbacks for that operation. Also added
a new Kotlin-only version `syncPurchasesWith` to provide an easier
Kotlin API.

Additionally, while doing this, I decided to move the logic from
`syncPurchases` to a new `SyncPurchasesHelper`, so it can be better
tested in isolation.