Refactor offerings code out of Purchases (#1027)
### Description
This is a first PR in preparation of persisting offerings in cache
(SDK-3094). There are no behavior changes in this PR. This just extracts
most of the code related to offerings from `Purchases` to the new
classes `OfferingsManager` and `OfferingsFactory`.

#### Main changes
- Created `OfferingsManager` that will be the entry point for all
offerings-related logic
- Created `OfferingsFactory` that allows to create the `Offerings`
object from a JSON response. It will query the store to get the
necessary products and create the final object.
- Moved `GoogleOfferingParser` from the `feature/google` module to the
`common` module. This is to be able to use the actual code in the unit
tests more easily.