Get offerings response from disk cache if available (#1029)
### Description
Third and final part of persisting offerings in disk to make the system
more robust. In this final PR, we use the offerings response cached in
#1028 so, in case of a backend error getting offerings, we can calculate
the offerings (as long as Play store/Amazon store still give us the
product information) as long as we have cached the value before.