PurchaseTester: Add new UI to configure internal proxy behavior (#1016)
### Description
This PR adds a new button to the `ConfigureFragment`, `LoginFragment`
and `OverviewFragment`. This will open a bottom sheet where we can
modify the behavior of our internal proxy. This will help with testing
offline entitlements and trusted entitlements. Note that this UI will
only be helpful when using our internal proxy. When using other proxies,
the requests will fail and the UI will display the corresponding error


https://github.com/RevenueCat/purchases-android/assets/808417/4455da06-0c44-40e2-b3a3-f0aa572f294b