Create merged report in purchases
The recommendation from Kover (https://github.com/Kotlin/kotlinx-kover/tree/v0.7.0/examples/android/multiproject) is that we do this merging in one module so purchases seems to make sense for this. The problem with this is that we'll need to remember to add any new module to the list. I did attempt to try this in other ways such as applying on the root, using other Kover attributes but this was the only successful way. Maybe I lack the proper Gradle fu.