Add plugins to toml file
We set all plugins in the base gradle file but we only apply the one's we're actually using in the root build.gradle. We can then use the aliases in other build.gradle and actually apply them where they are actually used.

The pluginManagement block is set the repositories from which we'll fetch the plugin dependencies