Fix billing connection error when querying purchases early in the process lifetime (#1032)
### Description
With the addition of offline entitlements, we query active purchases
early in the process lifetime. This sometimes resulted in the billing
client connection not being established at that point in time and not
calculating offline entitlements (returning an error).

This PR fixes those errors by making the queryPurchases method execute
AFTER the billing client connection has been established. We already do
this in all other methods that interact with the billing client except
this one, so it seemed like a bug.