[EXTERNAL] Update Kover to 0.7.0 (#1031)
### Checklist
- [ ] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation
~Resolves #356 by adding test coverage back.~ Code coverage was added
back in a while ago! This PR updates Kover to 0.7.0 and makes the
changes necessary to merge reports.

### Description

Kover has recently updated to 0.7.0, so the update was also performed as
part of this PR.

It required some changes to how Kover is set up and how the reports are
merged. I followed the advice in the [migration
guide](https://github.com/Kotlin/kotlinx-kover/blob/v0.7.0/docs/gradle-plugin/migrations/migration-to-0.7.0.md)
and the [android multiproject
example](https://github.com/Kotlin/kotlinx-kover/tree/v0.7.0/examples/android/multiproject).

I also updated the CircleCI script and added the test coverage badge
back to the readme.

Obviously, I don't have access to your CI to test this completely, but I
have run everything locally to check that it works.