Fix backend integration test flakiness (#1053)
### Description
I noticed a lot of flakiness around the backend integration tests. Looks
like it's because those were using the long jittering durations for the
requests, which can delay the request up to 10s. I changed to execute
the backend integration tests only in the integration test flavor which
reduces that amount to a maximum of 1s.