Refactor syncing pending transactions logic out of `Purchases` (#1058)
### Description
SDK-3155

This is another refactor in the process of posting unsynced purchases
before doing customer info requests, to make sure the information
received from the backend is up to date.

In this PR, we are extracting the logic that syncs purchases on app
foreground out of `Purchases` and added a callback to that process. The
idea is to reuse the same logic when trying to fetch customer info
first, and if that returns an empty customer info (that is, no unsynced
purchases), then we actually fetch customer info from the backend.