Sync pending purchases before getting customer info (#1073)
### Description
Followup to #1052 and #1058 

This changes the behavior of fetching customer info. Now, when fetching
customer info, we will first try to sync any pending purchases if any.
If there are pending purchases, we will return the customer info from
the last request sent. If there aren't, we will then hit the current
customer info endpoint to get the latest customer info.