Compose sample app (#1056)
### Description
This PR adds a new sample app `MagicWeatherCompose`, very similar to the
existing `MagicWeather` but using Jetpack Compose. Many things are
copied over from `MagicWeather` including the actual weather change
logic and the purchase SDK setup. It does change how we interact with
the SDK from there.

To better understand the code, this app has 2 main screens:
- Main
- Paywall

The `Main` screen has a bottom navigation view and has 2 subscreens,
`Weather` and `User`.

This can be a better sample for more modern android development. Things
remaining:
- [ ] Build app on every PR, to make sure it compiles.
- [ ] Improve UI
- [ ] Use coroutines
- [ ] Add some global state instead of always performing
`getCustomerInfo`, so we can use it to listen to the customer info
listener.


https://github.com/RevenueCat/purchases-android/assets/808417/7c7c54a9-a331-4e00-8329-0c5e86f2b1eb