Default customer info schema version to latest known by SDK (#1080)
### Description
An integration test started failing due to #1073. The error was that the
schemaVersion field in the customer info didn't match between the
fetched and cached customer info. It was actually luck that made it work
before 😬 .

Before that change, this test tried to get customer info twice: On sdk
configuration and the first part of the test. Since both of these happen
pretty quickly, we only made the request once and called both callbacks
with the same response. However, the first request actually modified the
JSONObject, adding the schema version and some other fields in order to
cache it in the device. This made it so the JSONObject actually had
those fields in the second callback, causing the schemaVersion to have a
value. Then, the value returned from cache already has the schemaVersion
field set so the test passed.

After the change in
https://github.com/RevenueCat/purchases-android/pull/1073, the test
still tries to get customer info twice, but the raw json sharing didn't
happen, since when we call getCustomerInfo we now need to check first
for unsynced purchases, causing us to not share the callback, so the
schema version that was set in the first callback wasn't shared with the
second one, causing the schemaVersion field to be 0, causing the test to
fail.

After talking about it, we think defaulting the schema version to the
latest known version by the SDK is probably expected behavior, instead
of defaulting to 0 before it's been saved.

Followup to #1078