Fix flaky integration tests: Wait for coroutines to finish before continuing (#1120)
### Description
Looks like our new coroutine test infrastructure added in #1077 / #1107
wasn't waiting for the code to be executed before continuing. In some
cases the `@After` code was actually executed before the test, in those
cases, the test was failing. In this PR I change it to a blocking method
so the test finishes before continuing.