`Trusted Entitlements`: stable (#1105)
This PR brings Trusted entitlements to stable and to java. 

#### Trusted entitlements

This new feature prevents MitM attacks between the SDK and the
RevenueCat server.
With verification enabled, the SDK ensures that the response created by
the server was not modified by a third-party, and the entitlements
received are exactly what was sent.
This is 100% opt-in. EntitlementInfos have a new VerificationResult
property, which will indicate the validity of the responses when this
feature is enabled.

```
fun configureRevenueCat() {
    val configuration = PurchasesConfiguration.Builder(context, apiKey)
        .entitlementVerificationMode(EntitlementVerificationMode.INFORMATIONAL)
        .build()
    Purchases.configure(configuration)
}
```

---------

Co-authored-by: Toni Rico <antonio.rico.diez@revenuecat.com>