Add `customEntitlementComputation` flavor (#1147)
### Description
We are adding a new flavor to separate the APIs needed for the
customEntitlementComputation package. In this PR, nothing will be
different between both. We will be running integration tests only in the
defaults flavor and we will be releasing only the defaults flavor. We
will start adding different functionality in future PRs.