CustomEntitlementComputation: New Purchases class (#1153)
### Description
This PR moves the existing `Purchases` class into the `defaults` flavor
and creates a new `Purchases` class in the
`customEntitlementComputation` flavor. It also moves some of the
extension functions to the `defaults` flavor since they aren't possible
in the `customEntitlementComputation` flavor. It also splits the tests
and api tests to run some tests on the `defaults` flavor only