Support receiving productId with basePlanId in upgrades/downgrades (#1183)
### Description
This is a possible approach to some issues we've had with developers
passing the wrong productId to the purchase methods on
upgrades/downgrades. We expect them to pass the productId, but some are
passing the `productId:basePlanId` form.

This PR supports both forms, by just picking the `productId` when given
in the incorrect form.