Remove customEntitlementComputation flavor for non-purchases modules (#1180)
### Description
This reverts the changes in #1169, and additionally makes some changes
so we only apply the maven publish plugin when using the `defaults`
flavor in non-purchases modules. This should avoid having to have extra
flavors in those modules and having them support the
`customEntitlementComputation` flavor.

This was becoming a bit of a problem as I was working in the debug view
#1075. Since some of the functions I'm using aren't in the
`customEntitlementComputation` flavor. I could have split the debug view
library by flavor to make it compatible, but that seems like unnecessary
work.